Split Kafka adapter tests into modules
