+++ test/components/bot_detector/event_queue/adapter_kafka/test_consumer_adapter.py
@@ -1,23 +1,16 @@
 from unittest.mock import AsyncMock, patch
 
-import orjson
 import pytest
-from aiokafka.errors import KafkaTimeoutError
 from bot_detector.event_queue.adapters.kafka import (
     AIOKafkaConsumerAdapter,
-    AIOKafkaAdapter,
-    AIOKafkaProducerAdapter,
     KafkaConfig,
     KafkaConsumerConfig,
-    KafkaProducerConfig,
 )
 from bot_detector.event_queue.core.errors import (
     ConsumerConfigError,
     ConsumerFetchError,
     ConsumerNotStartedError,
     MessageTypeError,
-    ProducerConfigError,
-    ProducerNotStartedError,
 )
 from pydantic import BaseModel, ValidationError
 
@@ -26,166 +19,6 @@ class PlayerScraped(BaseModel):
     id: int
     username: str
     score: int
-
-
-@pytest.mark.asyncio
-async def test_producer_start_stop():
-    config = KafkaConfig(
-        topic="players",
-        bootstrap_servers="localhost:9092",
-        consumer=False,
-        producer=True,
-        consumer_config=None,
-        producer_config=KafkaProducerConfig(partition_key_fn=lambda: "1"),
-    )
-    adapter = AIOKafkaProducerAdapter(PlayerScraped, config)
-
-    with patch(
-        "bot_detector.event_queue.adapters.kafka.adapter.AIOKafkaProducer"
-    ) as mock_producer_class:
-        mock_producer = mock_producer_class.return_value
-        mock_producer.start = AsyncMock()
-        mock_producer.stop = AsyncMock()
-
-        await adapter.start()
-        await adapter.stop()
-
-        mock_producer.start.assert_awaited_once()
-        mock_producer.stop.assert_awaited_once()
-
-
-@pytest.mark.asyncio
-async def test_producer_custom_partition_key():
-    config = KafkaConfig(
-        topic="players",
-        bootstrap_servers="localhost:9092",
-        consumer=False,
-        producer=True,
-        consumer_config=None,
-        producer_config=KafkaProducerConfig(partition_key_fn=lambda: "1"),
-    )
-    adapter = AIOKafkaProducerAdapter(PlayerScraped, config)
-
-    with patch(
-        "bot_detector.event_queue.adapters.kafka.adapter.AIOKafkaProducer"
-    ) as mock_producer_class:
-        mock_producer = mock_producer_class.return_value
-        mock_producer.start = AsyncMock()
-        mock_producer.send = AsyncMock()
-        mock_producer.stop = AsyncMock()
-
-        await adapter.start()
-        await adapter.put([PlayerScraped(id=9, username="Alice", score=100)])
-
-        expected_key = b"1"
-        mock_producer.send.assert_called_once_with(
-            topic="players",
-            value=PlayerScraped(id=9, username="Alice", score=100).model_dump(),
-            key=expected_key,
-        )
-
-
-@pytest.mark.asyncio
-async def test_producer_partition_key_bytes():
-    config = KafkaConfig(
-        topic="players",
-        bootstrap_servers="localhost:9092",
-        consumer=False,
-        producer=True,
-        consumer_config=None,
-        producer_config=KafkaProducerConfig(partition_key_fn=lambda: b"bytes"),
-    )
-    adapter = AIOKafkaProducerAdapter(PlayerScraped, config)
-
-    adapter.producer = AsyncMock()
-    adapter.producer.send = AsyncMock()
-
-    await adapter.put([PlayerScraped(id=9, username="Alice", score=100)])
-
-    adapter.producer.send.assert_called_once_with(
-        topic="players",
-        value=PlayerScraped(id=9, username="Alice", score=100).model_dump(),
-        key=b"bytes",
-    )
-
-
-@pytest.mark.asyncio
-async def test_producer_partition_key_invalid_type():
-    config = KafkaConfig(
-        topic="players",
-        bootstrap_servers="localhost:9092",
-        consumer=False,
-        producer=True,
-        consumer_config=None,
-        producer_config=KafkaProducerConfig(partition_key_fn=lambda: 123),
-    )
-    adapter = AIOKafkaProducerAdapter(PlayerScraped, config)
-    adapter.producer = AsyncMock()
-    adapter.producer.send = AsyncMock()
-
-    with pytest.raises(ValueError, match="partition_key_fn must return bytes or str"):
-        await adapter.put([PlayerScraped(id=9, username="Alice", score=100)])
-
-
-@pytest.mark.asyncio
-async def test_producer_serialization():
-    config = KafkaConfig(
-        topic="players",
-        bootstrap_servers="localhost:9092",
-        consumer=False,
-        producer=True,
-        consumer_config=None,
-        producer_config=KafkaProducerConfig(partition_key_fn=lambda: "1"),
-    )
-    adapter = AIOKafkaProducerAdapter(PlayerScraped, config)
-
-    with patch(
-        "bot_detector.event_queue.adapters.kafka.adapter.AIOKafkaProducer"
-    ) as mock_producer_class:
-        mock_producer = mock_producer_class.return_value
-        mock_producer.start = AsyncMock()
-
-        await adapter.start()
-
-        serializer = mock_producer_class.call_args.kwargs["value_serializer"]
-        payload = PlayerScraped(id=1, username="Alice", score=100).model_dump()
-        assert serializer(payload) == orjson.dumps(payload)
-
-
-@pytest.mark.asyncio
-async def test_producer_start_noop_when_existing():
-    config = KafkaConfig(
-        topic="players",
-        bootstrap_servers="localhost:9092",
-        consumer=False,
-        producer=True,
-        consumer_config=None,
-        producer_config=KafkaProducerConfig(partition_key_fn=lambda: "1"),
-    )
-    adapter = AIOKafkaProducerAdapter(PlayerScraped, config)
-    adapter.producer = AsyncMock()
-
-    with patch(
-        "bot_detector.event_queue.adapters.kafka.adapter.AIOKafkaProducer"
-    ) as mock_producer_class:
-        await adapter.start()
-
-        mock_producer_class.assert_not_called()
-
-
-@pytest.mark.asyncio
-async def test_producer_stop_noop_when_missing():
-    config = KafkaConfig(
-        topic="players",
-        bootstrap_servers="localhost:9092",
-        consumer=False,
-        producer=True,
-        consumer_config=None,
-        producer_config=KafkaProducerConfig(partition_key_fn=lambda: "1"),
-    )
-    adapter = AIOKafkaProducerAdapter(PlayerScraped, config)
-
-    await adapter.stop()
 
 
 @pytest.mark.asyncio
@@ -428,41 +261,6 @@ async def test_consumer_get_many_empty_records_sleeps():
 
 
 @pytest.mark.asyncio
-async def test_producer_send_failure():
-    config = KafkaConfig(
-        topic="players",
-        bootstrap_servers="localhost:9092",
-        consumer=False,
-        producer=True,
-        consumer_config=None,
-        producer_config=KafkaProducerConfig(
-            partition_key_fn=lambda: "1",
-            MAX_PRODUCE_RETRIES=1,
-        ),
-    )
-    adapter = AIOKafkaProducerAdapter(PlayerScraped, config)
-
-    with (
-        patch(
-            "bot_detector.event_queue.adapters.kafka.adapter.AIOKafkaProducer"
-        ) as mock_producer_class,
-        patch(
-            "bot_detector.event_queue.adapters.kafka.adapter.asyncio.sleep",
-            new=AsyncMock(),
-        ),
-    ):
-        mock_producer = mock_producer_class.return_value
-        mock_producer.start = AsyncMock()
-        mock_producer.send = AsyncMock(side_effect=KafkaTimeoutError("fail"))
-        mock_producer.stop = AsyncMock()
-
-        await adapter.start()
-        result = await adapter.put([PlayerScraped(id=1, username="Alice", score=100)])
-
-        assert isinstance(result, KafkaTimeoutError)
-
-
-@pytest.mark.asyncio
 async def test_consumer_connection_error():
     config = KafkaConfig(
         topic="players",
@@ -480,23 +278,6 @@ async def test_consumer_connection_error():
     result = await adapter.get_one()
     assert isinstance(result, ConsumerFetchError)
     assert isinstance(result.cause, Exception)
-
-
-@pytest.mark.asyncio
-async def test_producer_put_without_start():
-    config = KafkaConfig(
-        topic="players",
-        bootstrap_servers="localhost:9092",
-        consumer=False,
-        producer=True,
-        consumer_config=None,
-        producer_config=KafkaProducerConfig(partition_key_fn=lambda: "1"),
-    )
-    adapter = AIOKafkaProducerAdapter(PlayerScraped, config)
-
-    result = await adapter.put([PlayerScraped(id=1, username="Alice", score=100)])
-
-    assert isinstance(result, ProducerNotStartedError)
 
 
 @pytest.mark.asyncio
@@ -537,24 +318,6 @@ async def test_consumer_invalid_message_type():
     result = await adapter.get_one()
 
     assert isinstance(result, MessageTypeError)
-
-
-@pytest.mark.asyncio
-async def test_producer_put_without_config():
-    config = KafkaConfig.model_construct(
-        topic="players",
-        bootstrap_servers="localhost:9092",
-        consumer=False,
-        producer=True,
-        consumer_config=None,
-        producer_config=None,
-    )
-    adapter = AIOKafkaProducerAdapter(PlayerScraped, config)
-    adapter.producer = AsyncMock()
-
-    result = await adapter.put([PlayerScraped(id=1, username="Alice", score=100)])
-
-    assert isinstance(result, ProducerConfigError)
 
 
 @pytest.mark.asyncio
@@ -668,42 +431,3 @@ async def test_consumer_commit_success():
     await adapter.commit()
 
     adapter.consumer.commit.assert_awaited_once()
-
-
-@pytest.mark.asyncio
-async def test_queue_adapter_wiring():
-    config = KafkaConfig(
-        topic="players",
-        bootstrap_servers="localhost:9092",
-        consumer=True,
-        producer=True,
-        consumer_config=KafkaConsumerConfig(group_id="group"),
-        producer_config=KafkaProducerConfig(partition_key_fn=lambda: "1"),
-    )
-    adapter = AIOKafkaAdapter(PlayerScraped, config)
-    adapter.producer.start = AsyncMock()
-    adapter.consumer.start = AsyncMock()
-    adapter.producer.stop = AsyncMock()
-    adapter.consumer.stop = AsyncMock()
-    adapter.producer.put = AsyncMock()
-    adapter.consumer.get_one = AsyncMock(return_value="one")
-    adapter.consumer.get_many = AsyncMock(return_value=["many"])
-    adapter.consumer.commit = AsyncMock()
-
-    await adapter.start()
-    await adapter.put([PlayerScraped(id=1, username="Alice", score=100)])
-    one = await adapter.get_one()
-    many = await adapter.get_many(2)
-    await adapter.commit()
-    await adapter.stop()
-
-    adapter.producer.start.assert_awaited_once()
-    adapter.consumer.start.assert_awaited_once()
-    adapter.producer.put.assert_awaited_once()
-    adapter.consumer.get_one.assert_awaited_once()
-    adapter.consumer.get_many.assert_awaited_once()
-    adapter.consumer.commit.assert_awaited_once()
-    adapter.consumer.stop.assert_awaited_once()
-    adapter.producer.stop.assert_awaited_once()
-    assert one == "one"
-    assert many == ["many"]
